+++ drivers/input/Kconfig
@@ -7,6 +7,7 @@ menu "Input drivers"
 
 source "drivers/input/Kconfig.ft5336"
 source "drivers/input/Kconfig.gpio_keys"
+source "drivers/input/Kconfig.sdl"
 
 endmenu # Input Drivers
