+++ dts/bindings/input/zephyr,input-sdl-touch.yaml
@@ -1,0 +1,6 @@
+# Copyright (c) 2022 Kumar Gala <galak@kernel.org>
+# SPDX-License-Identifier: Apache-2.0
+
+description: SDL based emulated touch panel
+
+compatible: "zephyr,input-sdl-touch"
