--- dts/bindings/kscan/zephyr,sdl-kscan.yaml
@@ -1,8 +1,0 @@
-# Copyright (c) 2022 Kumar Gala <galak@kernel.org>
-# SPDX-License-Identifier: Apache-2.0
-
-description: SDL based emulated keyboard matrix devices
-
-compatible: "zephyr,sdl-kscan"
-
-include: kscan.yaml
